refactor: PromptGenerator 위치 변경 및 gradle 주석 추가
