+++ src/main/java/com/dawn/cs_study/ai/application/RagUseCase.java
@@ -4,7 +4,7 @@ import com.dawn.cs_study.ai.domain.DocumentSearchResult;
 import com.dawn.cs_study.ai.domain.QueryResult;
 import com.dawn.cs_study.ai.application.port.QueryAiPort;
 import com.dawn.cs_study.ai.application.port.VectorQueryPort;
-import com.dawn.cs_study.ai.domain.PromptGenerator;
+import com.dawn.cs_study.ai.domain.support.PromptGenerator;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.ai.chat.model.ChatResponse;
